+++ frontend/src/test/components/read/ReadingPage/Reply/ReadReply/ReadReReply.test.tsx
@@ -4,12 +4,12 @@ import { rest } from 'msw';
 import { setupServer } from 'msw/node';
 import { RecoilRoot } from 'recoil';
 
-import ReadReReply from '../../components/ReadReReply';
-import { refreshState } from '../../Recoil';
-import { onChange, RecoilObserver } from '../jest/RecoilObserver';
+import ReadReReply from '../../../../../../components/read/ReadingPage/Reply/ReadReply/ReadReReply';
+import { refreshState } from '../../../../../../Recoil';
+import { onChange, RecoilObserver } from '../../../../../jest/RecoilObserver';
 
 jest.mock(
-  '../../components/PutReply',
+  '../../../../../../components/read/ReadingPage/Reply/ReadReply/PutReply',
   () =>
     function () {
       return <div data-testid="PutReply" />;
